+++ src/components/rocm/roc_profiler.h
@@ -15,13 +15,13 @@ int rocp_init(void);
 int rocp_shutdown(void);
 
 /* native event interfaces */
-int rocp_evt_enum(unsigned int *event_code, int modifier);
-int rocp_evt_code_to_descr(unsigned int event_code, char *descr, int len);
-int rocp_evt_name_to_code(const char *name, unsigned int *event_code);
-int rocp_evt_code_to_name(unsigned int event_code, char *name, int len);
+int rocp_evt_enum(uint64_t *event_code, int modifier);
+int rocp_evt_code_to_descr(uint64_t event_code, char *descr, int len);
+int rocp_evt_name_to_code(const char *name, uint64_t *event_code);
+int rocp_evt_code_to_name(uint64_t event_code, char *name, int len);
 
 /* profiling context handling interfaces */
-int rocp_ctx_open(unsigned int *events_id, int num_events, rocp_ctx_t *ctx);
+int rocp_ctx_open(uint64_t *events_id, int num_events, rocp_ctx_t *ctx);
 int rocp_ctx_close(rocp_ctx_t ctx);
 int rocp_ctx_start(rocp_ctx_t ctx);
 int rocp_ctx_stop(rocp_ctx_t ctx);
